refactor(dapp): renamed team-payment-type enum to team-payment
