+++ apps/dapp/src/components/Pages/TeamPayments.tsx
@@ -11,7 +11,7 @@ import { TempleTeamPayments__factory } from 'types/typechain';
 import {
   TEAM_PAYMENTS_EPOCHS,
   TEAM_PAYMENTS_FIXED_ADDRESSES_BY_EPOCH,
-} from 'enums/team-payment-type';
+} from 'enums/team-payment';
 import { fromAtto } from 'utils/bigNumber';
 import withWallet from 'hoc/withWallet';
 import { useWallet } from 'providers/WalletProvider';
